+++ codegen.yml
@@ -1,8 +1,8 @@
-schema: './src/modules/**/schema.ts'
+schema: './src/**/schema.ts'
 generates:
-    './src/types/resolvers.ts':
+    './src/common/resolvers.types.ts':
         config:
-            contextType: './context#Context'
+            contextType: './types#Context'
             mappers:
                 Album: '../modules/albums#AlbumResponse'
                 Artist: '../modules/artists#ArtistResponse'
