+++ src/common/resolvers.types.ts
@@ -6,7 +6,7 @@ import { FavouritesResponse } from '../modules/favourites';
 import { GenreResponse } from '../modules/genres';
 import { TrackResponse } from '../modules/tracks';
 import { UserResponse } from '../modules/users';
-import { Context } from './context';
+import { Context } from './types';
 export type Maybe<T> = T | null;
 export type InputMaybe<T> = Maybe<T>;
 export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
