+++ src/common/types.ts
@@ -6,3 +6,10 @@ export type Context = {
     token: string
     dataSources: DataSources
 }
+
+export type Id = string
+
+export type Pagination = {
+    offset?: any
+    limit?: any
+}
