+++ src/modules/albums/service.ts
@@ -1,5 +1,5 @@
 import { dotEnvConfig } from '../../config'
-import { Service } from '../service'
+import { Service } from '../../common'
 import { AlbumResponse } from './types'
 
 export class AlbumsService extends Service<AlbumResponse> {
