+++ src/modules/bands/service.ts
@@ -1,5 +1,5 @@
 import { dotEnvConfig } from '../../config'
-import { Service } from '../service'
+import { Service } from '../../common'
 import { BandResponse } from './types'
 
 export class BandsService extends Service<BandResponse> {
