+++ src/modules/genres/service.ts
@@ -1,5 +1,5 @@
 import { dotEnvConfig } from '../../config'
-import { Service } from '../service'
+import { Service } from '../../common'
 import { GenreResponse } from './types'
 
 export class GenresService extends Service<GenreResponse> {
