+++ src/modules/tracks/service.ts
@@ -1,5 +1,5 @@
 import { dotEnvConfig } from '../../config'
-import { Service } from '../service'
+import { Service } from '../../common'
 import { TrackResponse } from './types'
 
 export class TracksService extends Service<TrackResponse> {
