--- src/types/index.ts
@@ -1,4 +1,0 @@
-export * from './context'
-export * from './pagination'
-export * from './resolvers'
-export * from './service'
