--- src/types/pagination.ts
@@ -1,4 +1,0 @@
-export type Pagination = {
-    offset?: any
-    limit?: any
-}
