--- src/types/service.ts
@@ -1,6 +1,0 @@
-export type Id = string
-
-export type DeleteResponse = {
-    acknowledged: boolean
-    deletedCount: number
-}
